Clean sun vector tests
- Moved sun vector tests to test file
- Currently, median 0.6º error as compared to ephemeris
