+++ src/tests/sun_vector_tests.h
@@ -1,0 +1,13 @@
+/**
+ * @author Chen Li and Lundeen Cahilly
+ * @date 2025-09-05
+ *
+ * Sun vector model tests
+ */
+
+#pragma once
+
+#include "slate.h"
+
+void test_sun_vector_eci(slate_t *slate);
+void test_sun_vector_year(slate_t *slate);
